requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiomysql==0.2.0
aioresponses==0.7.7
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.13.3
alluka==0.3.3
annotated-types==0.7.0
apscheduler==3.11.0
asyncpg==0.30.0
attrs==25.1.0
beautifulsoup4==4.13.3
brotli==1.1.0
cairocffi==1.7.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
ciso8601==2.3.2
click==8.1.7
colorlog==6.9.0
cryptography==43.0.3
cssselect2==0.7.0
defusedxml==0.7.1
dnspython==2.7.0
frozenlist==1.5.0
gitdb==4.0.11
gitpython==3.1.43
greenlet==3.1.1
hikari==2.1.1
hikari-arc==1.4.0
hikari-miru==4.2.0
idna==3.10
iniconfig==2.0.0
levenshtein==0.26.1
loguru==0.7.2
lxml==5.3.0
mako==1.3.6
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
motor==3.6.0
multidict==6.1.0
munch==4.0.0
nodeenv==1.9.1
orjson==3.10.15
packaging==24.2
pillow==11.0.0
pip==24.3.1
pluggy==1.5.0
propcache==0.3.0
psutil==6.1.0
pycares==4.5.0
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pygit==0.1
pygments==2.18.0
pymongo==4.9.2
pymysql==1.1.1
pyright==1.1.396
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
pytz==2024.2
rapidfuzz==3.12.1
reportlab==4.2.5
requests==2.32.3
rich==13.9.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
soupsieve==2.6
sqlalchemy==2.0.36
tinycss2==1.4.0
typer==0.12.5
types-cffi==1.16.0.20240331
types-colorama==0.4.15.20240311
types-croniter==3.0.4.20241027
types-pyopenssl==24.1.0.20240722
types-requests==2.32.0.20241016
types-setuptools==75.2.0.20241025
typing-extensions==4.12.2
tzlocal==5.3
urllib3==2.2.3
webencodings==0.5.1
yarl==1.18.3
